move constants to own module
